requirements.txt
# Core scientific stack (adjust as needed)
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
matplotlib==3.9.0
seaborn==0.13.2
PyYAML==6.0.2

